+++ packages/orpheus/android/src/main/java/expo/modules/orpheus/service/OrpheusMusicService.kt
@@ -93,7 +93,6 @@ class OrpheusMusicService : MediaLibraryService() {
 
         GeneralStorage.initialize(this)
         LoudnessStorage.initialize(this)
-        expo.modules.orpheus.network.CronetManager.init(this)
 
         setMediaNotificationProvider(object : DefaultMediaNotificationProvider(this) {
             override fun getMediaButtons(